Moved RestoreOptions_* Classes to new Seperate Class File
