+++ PSAsigraDSClient/NewDSClientFSRestoreOptions.cs
@@ -1,4 +1,5 @@
 ﻿using System.Management.Automation;
+using static PSAsigraDSClient.DSClientRestoreOptions;
 
 namespace PSAsigraDSClient
 {
